+++ tests/test_simulatedorder.py
@@ -162,7 +162,7 @@ class SimulatedOrderTest(unittest.TestCase):
             self.simulated.place(mock_order_package, mock_market_book, {}, 1)
 
     @mock.patch("flumine.simulation.simulatedorder.SimulatedOrder._get_runner")
-    def test_place_limit_back_fill_or_kill(self, mock__get_runner):
+    def test_place_limit_back_fill_or_kill_matched(self, mock__get_runner):
         mock_client = mock.Mock(best_price_execution=True)
         mock_order_package = mock.Mock(client=mock_client, market_version=None)
         mock_market_book = mock.Mock(status="OPEN")
@@ -175,12 +175,34 @@ class SimulatedOrderTest(unittest.TestCase):
             mock_order_package, mock_market_book, instruction, 1
         )
         self.assertEqual(self.simulated.market_version, mock_market_book.version)
+        self.assertEqual(resp.average_price_matched, 12)
+        self.assertEqual(resp.size_matched, 1)
+        self.assertEqual(
+            self.simulated.matched, [[mock_market_book.publish_time_epoch, 12, 1]]
+        )
+
+    @mock.patch("flumine.simulation.simulatedorder.SimulatedOrder._get_runner")
+    def test_place_limit_back_fill_or_kill_cancelled(self, mock__get_runner):
+        mock_client = mock.Mock(best_price_execution=True)
+        mock_order_package = mock.Mock(client=mock_client, market_version=None)
+        mock_market_book = mock.Mock(status="OPEN")
+        mock_runner = mock.Mock()
+        mock_runner.ex.available_to_back = [{"price": 11, "size": 1}]
+        mock_runner.ex.available_to_lay = [{"price": 13, "size": 120}]
+        mock__get_runner.return_value = mock_runner
+        instruction = {"limitOrder": {"timeInForce": "FILL_OR_KILL"}}
+        resp = self.simulated.place(
+            mock_order_package, mock_market_book, instruction, 1
+        )
+        self.assertEqual(self.simulated.market_version, mock_market_book.version)
         self.assertEqual(resp.average_price_matched, 0)
         self.assertEqual(resp.size_matched, 0)
         self.assertEqual(self.simulated.matched, [])
 
     @mock.patch("flumine.simulation.simulatedorder.SimulatedOrder._get_runner")
-    def test_place_limit_back_fill_or_kill_min_fill_size(self, mock__get_runner):
+    def test_place_limit_back_fill_or_kill_min_fill_size_matched(
+        self, mock__get_runner
+    ):
         mock_client = mock.Mock(best_price_execution=True)
         mock_order_package = mock.Mock(client=mock_client, market_version=None)
         mock_market_book = mock.Mock(status="OPEN")
@@ -198,6 +220,28 @@ class SimulatedOrderTest(unittest.TestCase):
         self.assertEqual(
             self.simulated.matched, [[mock_market_book.publish_time_epoch, 12, 1]]
         )
+
+    @mock.patch("flumine.simulation.simulatedorder.SimulatedOrder._get_runner")
+    def test_place_limit_back_fill_or_kill_min_fill_size_cancelled(
+        self, mock__get_runner
+    ):
+        mock_client = mock.Mock(best_price_execution=True)
+        mock_order_package = mock.Mock(client=mock_client, market_version=None)
+        mock_market_book = mock.Mock(status="OPEN")
+        mock_runner = mock.Mock()
+        mock_runner.ex.available_to_back = [{"price": 12, "size": 1}]
+        mock_runner.ex.available_to_lay = [{"price": 13, "size": 120}]
+        mock__get_runner.return_value = mock_runner
+        instruction = {
+            "limitOrder": {"timeInForce": "FILL_OR_KILL", "minFillSize": 1.01}
+        }
+        resp = self.simulated.place(
+            mock_order_package, mock_market_book, instruction, 1
+        )
+        self.assertEqual(self.simulated.market_version, mock_market_book.version)
+        self.assertEqual(resp.average_price_matched, 0)
+        self.assertEqual(resp.size_matched, 0)
+        self.assertEqual(self.simulated.matched, [])
 
     @mock.patch("flumine.simulation.simulatedorder.SimulatedOrder._get_runner")
     def test_place_limit_back_fill_or_kill_vwap(self, mock__get_runner):
@@ -306,7 +350,7 @@ class SimulatedOrderTest(unittest.TestCase):
         )
 
     @mock.patch("flumine.simulation.simulatedorder.SimulatedOrder._get_runner")
-    def test_place_limit_lay_fill_or_kill(self, mock__get_runner):
+    def test_place_limit_lay_fill_or_kill_matched(self, mock__get_runner):
         mock_client = mock.Mock(best_price_execution=True)
         mock_order_package = mock.Mock(client=mock_client, market_version=None)
         self.simulated.order.side = "LAY"
@@ -314,6 +358,45 @@ class SimulatedOrderTest(unittest.TestCase):
         mock_runner = mock.Mock()
         mock_runner.ex.available_to_back = [{"price": 11, "size": 120}]
         mock_runner.ex.available_to_lay = [{"price": 12, "size": 1}]
+        mock__get_runner.return_value = mock_runner
+        instruction = {"limitOrder": {"timeInForce": "FILL_OR_KILL"}}
+        resp = self.simulated.place(
+            mock_order_package, mock_market_book, instruction, 1
+        )
+        self.assertEqual(resp.average_price_matched, 12)
+        self.assertEqual(resp.size_matched, 1)
+        self.assertEqual(
+            self.simulated.matched, [[mock_market_book.publish_time_epoch, 12, 1]]
+        )
+
+    @mock.patch("flumine.simulation.simulatedorder.SimulatedOrder._get_runner")
+    def test_place_limit_lay_fill_or_kill_cancelled(self, mock__get_runner):
+        mock_client = mock.Mock(best_price_execution=True)
+        mock_order_package = mock.Mock(client=mock_client, market_version=None)
+        self.simulated.order.side = "LAY"
+        mock_market_book = mock.Mock(status="OPEN")
+        mock_runner = mock.Mock()
+        mock_runner.ex.available_to_back = [{"price": 11, "size": 120}]
+        mock_runner.ex.available_to_lay = [{"price": 13, "size": 1}]
+        mock__get_runner.return_value = mock_runner
+        instruction = {"limitOrder": {"timeInForce": "FILL_OR_KILL"}}
+        resp = self.simulated.place(
+            mock_order_package, mock_market_book, instruction, 1
+        )
+        self.assertEqual(resp.average_price_matched, 0)
+        self.assertEqual(resp.size_matched, 0)
+        self.assertEqual(self.simulated.matched, [])
+
+    @mock.patch("flumine.simulation.simulatedorder.SimulatedOrder._get_runner")
+    def test_place_limit_lay_fill_or_kill_no_price(self, mock__get_runner):
+        self.mock_order.order_type.price = 1000
+        mock_client = mock.Mock(best_price_execution=True)
+        mock_order_package = mock.Mock(client=mock_client, market_version=None)
+        self.simulated.order.side = "LAY"
+        mock_market_book = mock.Mock(status="OPEN")
+        mock_runner = mock.Mock()
+        mock_runner.ex.available_to_back = []
+        mock_runner.ex.available_to_lay = []
         mock__get_runner.return_value = mock_runner
         instruction = {"limitOrder": {"timeInForce": "FILL_OR_KILL"}}
         resp = self.simulated.place(
@@ -342,6 +425,28 @@ class SimulatedOrderTest(unittest.TestCase):
         self.assertEqual(
             self.simulated.matched, [[mock_market_book.publish_time_epoch, 12, 1]]
         )
+
+    @mock.patch("flumine.simulation.simulatedorder.SimulatedOrder._get_runner")
+    def test_place_limit_lay_fill_or_kill_min_fill_size_cancelled(
+        self, mock__get_runner
+    ):
+        mock_client = mock.Mock(best_price_execution=True)
+        mock_order_package = mock.Mock(client=mock_client, market_version=None)
+        self.simulated.order.side = "LAY"
+        mock_market_book = mock.Mock(status="OPEN")
+        mock_runner = mock.Mock()
+        mock_runner.ex.available_to_back = [{"price": 11, "size": 120}]
+        mock_runner.ex.available_to_lay = [{"price": 12, "size": 1}]
+        mock__get_runner.return_value = mock_runner
+        instruction = {
+            "limitOrder": {"timeInForce": "FILL_OR_KILL", "minFillSize": 1.01}
+        }
+        resp = self.simulated.place(
+            mock_order_package, mock_market_book, instruction, 1
+        )
+        self.assertEqual(resp.average_price_matched, 0)
+        self.assertEqual(resp.size_matched, 0)
+        self.assertEqual(self.simulated.matched, [])
 
     @mock.patch("flumine.simulation.simulatedorder.SimulatedOrder._get_runner")
     def test_place_limit_lay_fill_or_kill_vwap(self, mock__get_runner):
@@ -889,8 +994,23 @@ class SimulatedOrderTest(unittest.TestCase):
         "flumine.simulation.simulatedorder.SimulatedOrder._calculate_process_traded"
     )
     def test__process_traded_back(self, mock__calculate_process_traded):
-        self.simulated._process_traded(1234579, {12: 120})
+        mock__calculate_process_traded.return_value = 5.0
+        traded = {12: 120}
+        self.simulated._process_traded(1234579, traded)
         mock__calculate_process_traded.assert_called_with(1234579, 120)
+        self.assertEqual(traded, {12: 115})
+
+    @mock.patch(
+        "flumine.simulation.simulatedorder.SimulatedOrder._calculate_process_traded"
+    )
+    def test__process_traded_over_fill(self, mock__calculate_process_traded):
+        for side in ("BACK", "LAY"):
+            self.simulated.order.side = side
+            mock__calculate_process_traded.return_value = 130.0
+            traded = {12: 120}
+            self.simulated._process_traded(1234579, traded)
+            mock__calculate_process_traded.assert_called_with(1234579, 120)
+            self.assertEqual(traded, {12: 0})
 
     @mock.patch(
         "flumine.simulation.simulatedorder.SimulatedOrder._calculate_process_traded"
@@ -903,9 +1023,12 @@ class SimulatedOrderTest(unittest.TestCase):
         "flumine.simulation.simulatedorder.SimulatedOrder._calculate_process_traded"
     )
     def test__process_traded_lay(self, mock__calculate_process_traded):
+        mock__calculate_process_traded.return_value = 10.0
         self.simulated.order.side = "LAY"
-        self.simulated._process_traded(1234581, {12: 120})
+        traded = {12: 120}
+        self.simulated._process_traded(1234581, traded)
         mock__calculate_process_traded.assert_called_with(1234581, 120)
+        self.assertEqual(traded, {12: 110})
 
     @mock.patch(
         "flumine.simulation.simulatedorder.SimulatedOrder._calculate_process_traded"